Feat: current holdings mostly done
